+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-03T14:05:30.205Z"
+    created: Optional[str] = "2026-07-03T16:56:46.999Z"
 
-    updated: Optional[str] = "2026-07-03T14:05:30.205Z"
+    updated: Optional[str] = "2026-07-03T16:56:46.999Z"
 
     updated_by_id: OptionalNullable[str] = UNSET