+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWM4SKNG8M3NQC2QAF7WFZS9"
+        "01KWMEK41MKPV6DPDWETNWCFSW"
     )
     r"""The unique identifier of the data source"""