+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWM4SKNK9S5E6W4XQAVY8CBQ"
+        "01KWMEK41P0G1RED74X23BV4KT"
     )
     r"""The unique identifier of the data source"""