+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWM4SKNSCAN1B564CYEPE015"
+        "01KWMEK41T4QEJAVV1E9KG4PFM"
     )
     r"""The unique identifier of the data source"""